Move the Backend to SQLite3 and Heavily Refactor
This commit began as a simple change: moving from flat text files to an
SQLite database. Within a few minutes of examining the code I knew there
was a major refactor job to do.

The huge Javascript file was spread out into seven smaller files.
Unnecessary functions were removed. Variable names were changed for
clearer intention. Console output was made more verbose to indicate when
things are changing on the server side.

With the code in a more readable state, I could implement the fairly
straightforward move to SQLite3. This required rewriting nearly all of
the PHP to communicate with a single database file and return
confirmation/error statements when doing so.

Lastly, I searched for a way to make the docker installation cleaner,
ie not having the chown the volume, but I never found anything useful.
The container has to run as root (apprently), so the volume it creates
is owned by root, and can't be changed within the Dockerfile. Perhaps in
the future I will look for a different container that doesn't run as
root, but for now this will have to do.